requirements.txt
fastapi~=0.115.12
uvicorn
requests~=2.32.3
pandas~=2.2.3
streamlit~=1.44.1

plotly~=6.0.1
django-scheduler~=0.10.1